requirements.txt
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.8.99
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.7
et-xmlfile==1.1.0
fire==0.5.0
Flask==3.0.0
flask-babel==4.0.0
fonttools==4.47.0
future==0.18.3
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
loguru==0.7.2
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.2
paddleocr==2.7.0.3
pandas==2.1.4
pdf2docx==0.5.7
pillow==10.2.0
premailer==3.10.0
protobuf==4.25.1
psutil==5.9.7
pyclipper==1.3.0.post5
pycryptodome==3.19.1
PyMuPDF==1.20.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
pywebio==1.8.3
PyYAML==6.0.1
rapidfuzz==3.6.1
rarfile==4.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
soupsieve==2.5
termcolor==2.4.0
tifffile==2023.12.9
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ua-parser==0.18.0
urllib3==2.1.0
user-agents==2.2.0
visualdl==2.5.3
Werkzeug==3.0.1
win32-setctime==1.1.0
zipp==3.17.0
